l>=mu and added K2=0, lam=lam0>0 case
Changed all sums over the Eigenvalue index, l, from stating at zero to
starting at mu.  This is because values lest then mu are not applicable.

Also added corner case when K0=0, and lam=lam0>0.  This is nearly the
same as the K0 > 0 expression except that there is no sum over l2 and
the only eignevalue to contribure is for l=lam at lam*(lam+D-2).